split pessimistic
