+++ testcase/pessimistic/pkg/pessimistic.go
@@ -1,4 +1,4 @@
-package pessimistic
+package pkg
 
 import (
 	"database/sql"
