+++ testcase/pessimistic/pkg/stmt.go
@@ -1,4 +1,4 @@
-package pessimistic
+package pkg
 
 import (
 	"context"
